Fix initialization semantics.
Before PR#1, a CA was considered 'initialized' when the CA certificate
had been created successfully. This was then changed to 'when the
directory exists', leading to issues when the directory was pre-created.
Unify these semantics so that existance of the folder, existance of the
pki subfolder and existance of the ca certificate are checked
separately.