Add unique certificate file distinguisher
Add an additional variable to incorporate into the file name for keys
and certificates.

This solves the problem of deploying several certificates of the same CA
to the same host which are required to carry the FQDN as CN. As of now
this results in duplication of the same certificate. This change allows
using different certificates in this scenario.